+++ src/container/GeoWebSystem/GWS.tsx
@@ -4,7 +4,7 @@ import TitleBar from "../../components/common/TitleBar/TitleBar";
 import GWLoader from "../../components/common/Loader/Loader";
 import GWAvail from "../../components/common/ContentFiller/Avail";
 import GWContentView from "../GeoWebInterface/components/GeoWebContent/GWContent";
-import { NETWORK_ID } from "../../lib/constants";
+import { WS_RPC_URL, NETWORK_ID } from "../../lib/constants";
 import { getGeoId, getParcelInfo } from "../../lib/api";
 import { ethers } from "ethers";
 import { useMUD } from "@geo-web/mud-world-base-client";
@@ -12,8 +12,61 @@ import { useComponentValue, useEntityQuery } from "@latticexyz/react";
 import { singletonEntity } from "@latticexyz/store-sync/recs";
 import { Has } from "@latticexyz/recs";
 import { useSearchParams } from "next/navigation";
+import {
+  MUDProvider,
+  SetupResult,
+  setup,
+} from "@geo-web/mud-world-base-client";
+import { optimismGoerli } from "viem/chains";
+import { MUDChain } from "@latticexyz/common/chains";
+
+const chainId = 420;
+const supportedChains: MUDChain[] = [
+  {
+    ...optimismGoerli,
+    rpcUrls: {
+      ...optimismGoerli.rpcUrls,
+      default: {
+        http: optimismGoerli.rpcUrls.default.http,
+        webSocket: [WS_RPC_URL],
+      },
+    },
+  },
+];
 
 export default function GWS() {
+  const params = useSearchParams();
+  const world = params.get("world");
+
+  const [mudSetup, setMUDSetup] = useState<SetupResult | null>(null);
+  useEffect(() => {
+    (async () => {
+      const mudSetup = await setup({
+        chainId,
+        worlds: {
+          [chainId]: {
+            address: world,
+            blockNumber: 14300510,
+          },
+        },
+        supportedChains,
+      });
+      setMUDSetup(mudSetup);
+    })();
+  }, []);
+
+  if (!mudSetup) {
+    return <GWLoader />;
+  }
+
+  return (
+    <MUDProvider value={mudSetup}>
+      <InnerGWS />
+    </MUDProvider>
+  );
+}
+
+function InnerGWS() {
   const params = useSearchParams();
   const lat = params.get("latitude");
   const lon = params.get("longitude");
