--- src/mud/worlds.json
@@ -1,6 +1,0 @@
-{
-  "420": {
-    "address": "0x5D59444Be22c322F0fe0076DaD760962213Ff9F2",
-    "blockNumber": 14300510
-  }
-}
